Extract event from the temporal worker run factory
